--- pytorch-qa/inference/demo/without_context/requirements.txt
@@ -1,7 +1,0 @@
-torch >= 2.0.0
-transformers
-langchain
-sentencepiece
-accelerate
-gradio
-huggingface-hub
